==> Capstone_01/00_erase_doc_name_datos_ambulatorios_ano.ipynb <==
## Relevant modules.
import os
import pandas as pd

files_in = ['original_data/datos_ambulatorios_2008.csv',\
            'original_data/datos_ambulatorios_2009.csv',\
            'original_data/datos_ambulatorios_2010.csv',\
            'original_data/datos_ambulatorios_2011.csv',\
            'original_data/datos_ambulatorios_2012.csv']

## Getting year of each file. 
## Reading each file as data_outpatient_year.
for file in files_in:
    year = file[-8:-4]
    data_outpatient_year = pd.read_csv(file, header=0)

    ## Creating new reference for data_outpatient_year['MEDICO'].
    ## New reference will omit the names of doctors.
    new_doc_name = []
    for i in data_outpatient_year['MEDICO']:
        name = list(i.split('-'))
        new_doc_name.append(name[0]+'-'+name[1]+'-'+'NOMBRE*DEL*DOCTOR')
        
    data_outpatient_year['MEDICO'] = new_doc_name
    
    ## Saving data_outpatient_year as csv to original_data folder.
    if os.path.exists('original_data/datos_ambulatorios_'+year+'_new.csv'):
        print('original_data/datos_ambulatorios_'+year+'_new.csv'+' already exists.')
    elif os.path.exists('original_data'):
        file_out = 'original_data/datos_ambulatorios_'+year+'_new.csv'
        data_outpatient_year.to_csv(path_or_buf=file_out, index=False)
        print('original_data/'+' already exists.')
        print('original_data/datos_ambulatorios_'+year+'_new.csv'+' has been created.')
    else:
        os.mkdir('original_data')
        file_out = 'original_data/datos_ambulatorios_'+year+'_new.csv'
        data_outpatient_year.to_csv(path_or_buf=file_out, index=False)
        print('original_data/datos_ambulatorios_'+year+'_new.csv'+' has been created.')
