==> Capstone_01/01_optional_data_outpatient_year_details.ipynb <==
## Relevant modules.
import pandas as pd
import os
## Optional for alternative code two (see forth cell below)
#import re

## First look at the original dataframe: data_outpatient_year_firstlook.
## Please choose a year between 2008-2012.
year = 2008

file_in = 'original_data/datos_ambulatorios_'+str(year)+'_new.csv'
data_outpatient_year_firstlook = pd.read_csv(file_in)
print(data_outpatient_year_firstlook.head())

## Organizing dataframe better: data_outpatient_year.
data_outpatient_year = pd.read_csv(file_in, header=0)
data_outpatient_year['year'] = [year]*len(data_outpatient_year)
col_names = ['doc_name','spec_code','spec_es','n_visits','n_surgeries','n_patients','year']
col_order = ['spec_code','spec_es','year','doc_name','n_visits','n_surgeries','n_patients']
data_outpatient_year.columns = col_names
data_outpatient_year = data_outpatient_year[col_order]

data_outpatient_year['spec_es'] = [i.lower() for i in data_outpatient_year['spec_es']]

print(data_outpatient_year.head(),'\n')
print(data_outpatient_year.info())
for i in range(len(data_outpatient_year.columns)):
    i_name = data_outpatient_year.iloc[:,i].name
    i_value = sum(list(type(i)==str for i in data_outpatient_year.iloc[:,i]))
    print('Column', i_name, 'has', i_value, 'strings.')

## Getting code for doctor, doc_code, and adding it to data_outpatient_year.
doc_code = []
for i in data_outpatient_year['doc_name']:
    for j in i.split('-'):
        if j.isdigit():
            doc_code.append(int(j))

## Alternative code one.
#doc_code = [int(j) for i in data_outpatient_year['doc_name'] for j in i.split('-') if j.isdigit()]            

## Alternative code two.
#doc_code = []
#for i in data_outpatient_year['doc_name']:
#    temp = re.findall(r'\d+', i) 
#    res = list(map(int, temp))
#    doc_code.append(res[0])

data_outpatient_year['doc_code'] = doc_code
col_order = ['spec_code','spec_es','year','doc_code','doc_name','n_visits','n_surgeries','n_patients']
data_outpatient_year = data_outpatient_year[col_order]

print(data_outpatient_year.head(),'\n')
print(data_outpatient_year.info())
for i in range(len(data_outpatient_year.columns)):
    i_name = data_outpatient_year.iloc[:,i].name
    i_value = sum(list(type(i)==str for i in data_outpatient_year.iloc[:,i]))
    print('Column', i_name, 'has', i_value, 'strings.')

## Sorting by spec_code and reseting index.
data_outpatient_year.drop('doc_name', axis=1, inplace=True)
data_outpatient_year.sort_values('spec_code', inplace=True)
data_outpatient_year.reset_index(drop=True, inplace=True)
print(data_outpatient_year.head(10),'\n')
