==> Capstone_01/03_data_departments.ipynb <==
## Relevant modules.
import pandas as pd
import os

## First look at the original dataframe: data_dep_firstlook.
file_in = 'original_data/datos_departamentos.csv'
data_dep_firstlook = pd.read_csv(file_in, encoding='latin-1')
print(data_dep_firstlook.head())

## Organizing dataframe better: data_dep.
data_dep = pd.read_csv(file_in, header=0, encoding='latin-1')
col_names = ['year','doc_code','spec_code','spec_es','dep_code','department']
col_order = ['spec_code','spec_es','year','doc_code','dep_code','department']
data_dep.columns = col_names
data_dep = data_dep[col_order]

data_dep['spec_es'] = [i.lower() for i in data_dep['spec_es']]
data_dep['department'] = [i.lower() for i in data_dep['department']]

print(data_dep.head(),'\n')
print(data_dep.info())
for i in range(len(data_dep.columns)):
    i_name = data_dep.iloc[:,i].name
    i_value = sum(list(type(i)==str for i in data_dep.iloc[:,i]))
    print('Column', i_name, 'has', i_value, 'strings.')

## Sorting by spec_code and reseting index.
data_dep.sort_values('spec_code', inplace=True)
data_dep.reset_index(drop=True, inplace=True)
print(data_dep.head(10),'\n')

## Saving data_dep as csv to data folder.
if os.path.exists('data/data_departments.csv'):
    print('data/data_departments.csv already exists.')
elif os.path.exists('data'):
    file_out = 'data/data_departments.csv'
    data_dep.to_csv(path_or_buf=file_out, index=False)
    print('data/ already exists.')
    print('data/data_departments.csv has been created.')
else:
    os.mkdir('data')
    file_out = 'data/data_departments.csv'
    data_dep.to_csv(path_or_buf=file_out, index=False)
    print('data/data_departments.csv has been created.')
