==> Capstone_01/04_panel.ipynb <==
## Relevant modules.
import pandas as pd
import os

## Reading files.
file_2008 = 'data/data_outpatient_2008.csv'
file_2009 = 'data/data_outpatient_2009.csv'
file_2010 = 'data/data_outpatient_2010.csv'
file_2011 = 'data/data_outpatient_2011.csv'
file_2012 = 'data/data_outpatient_2012.csv'
file_costs = 'data/mean_outpatient_costs_usd.csv'
file_dep = 'data/data_departments.csv'
data_outpatient_2008 = pd.read_csv(file_2008, header=0)
data_outpatient_2009 = pd.read_csv(file_2009, header=0)
data_outpatient_2010 = pd.read_csv(file_2010, header=0)
data_outpatient_2011 = pd.read_csv(file_2011, header=0)
data_outpatient_2012 = pd.read_csv(file_2012, header=0)
mean_outpatient_costs_usd = pd.read_csv(file_costs, header=0)
data_dep = pd.read_csv(file_dep, header=0)
## Expected length when concatenating 2008-2012 files. 
exp_len = (len(data_outpatient_2008)+
           len(data_outpatient_2009)+
           len(data_outpatient_2010)+
           len(data_outpatient_2011)+
           len(data_outpatient_2012))
## Expected total visits when concatenating 2008-2012 files. 
exp_vis = (data_outpatient_2008['n_visits'].sum()+
           data_outpatient_2009['n_visits'].sum()+
           data_outpatient_2010['n_visits'].sum()+
           data_outpatient_2011['n_visits'].sum()+
           data_outpatient_2012['n_visits'].sum())
## Expected total surgeries when concatenating 2008-2012 files. 
exp_sur = (data_outpatient_2008['n_surgeries'].sum()+
           data_outpatient_2009['n_surgeries'].sum()+
           data_outpatient_2010['n_surgeries'].sum()+
           data_outpatient_2011['n_surgeries'].sum()+
           data_outpatient_2012['n_surgeries'].sum())
## Expected total patients when concatenating 2008-2012 files. 
exp_pat = (data_outpatient_2008['n_patients'].sum()+
           data_outpatient_2009['n_patients'].sum()+
           data_outpatient_2010['n_patients'].sum()+
           data_outpatient_2011['n_patients'].sum()+
           data_outpatient_2012['n_patients'].sum())
print('The expected length when concatenating 2008-2012 files is', exp_len, end='.\n')
print('The expected total visits is', exp_vis, end='.\n')
print('The expected total surgeries is', exp_sur, end='.\n')
print('The expected total patients is', exp_pat, end='.')


## Concatenating 2008-2012 files.
concat_data = pd.concat([data_outpatient_2008,\
                         data_outpatient_2009,\
                         data_outpatient_2010,\
                         data_outpatient_2011,\
                         data_outpatient_2012],
                         ignore_index=True)

print(concat_data.head(), '\n')
print(concat_data.tail(), '\n')
print(concat_data.info(), '\n')
print('Length as expected?', len(concat_data) == exp_len, end='.') 

## Grouping by spec_code, spec_es, year and doc_code.
concat_data_unique = concat_data.groupby(['spec_code',\
                                          'spec_es',\
                                          'year',\
                                          'doc_code'],\
                                           as_index=False, sort=False).sum()

tot_vis = concat_data_unique['n_visits'].sum()
tot_sur = concat_data_unique['n_surgeries'].sum()
tot_pat = concat_data_unique['n_patients'].sum()

print(concat_data_unique.head(), '\n')
print(concat_data_unique.tail(), '\n')
print(concat_data_unique.info(), '\n')
print('Total visits is', tot_vis, end='.'); print(' As expected?', tot_vis==exp_vis, end='.\n')
print('Total surgeries is', tot_sur, end='.'); print(' As expected?', tot_sur==exp_sur, end='.\n')
print('Total patients is', tot_pat, end='.'); print(' As expected?', tot_pat==exp_pat, end='.')

## Putting everything together, a.k.a. merging.
panel_temp = pd.merge(concat_data_unique, mean_outpatient_costs_usd,\
                      on=['spec_code','spec_es','year'], how='left')
panel = pd.merge(panel_temp, data_dep,\
                 on=['spec_code','spec_es','year','doc_code'], how='left')

print(panel.head(), '\n')
print(panel.tail(), '\n')
print(panel.info())

## Making some rearrangement.
col_order = ['doc_code','spec_code','spec_es','spec_en',\
             'year','n_visits','n_surgeries','n_patients',\
             'c_visit','c_surgery','dep_code','department']
panel = panel[col_order]
print(panel.head(), '\n')
print(panel.tail(), '\n')
print(panel.info())

## Issue 01: errors with spec_es or spec_code.
filter_01 = panel['spec_en'].isnull()
print(panel[filter_01])

## Issue 02: no information about costs.
filter_02 = panel['c_visit'].isnull()
filter_03 = panel['c_surgery'].isnull()
print(panel[~filter_01 & (filter_02 | filter_03)])

## Saving data_panel without issue 01: spec_es or spec_code errors.
data_panel = panel[~filter_01]
data_panel.reset_index(drop=True, inplace=True)
print(data_panel.info(),'\n')
## Saving data_panel as csv to data folder.
if os.path.exists('data/data_panel.csv'):
    print('data/data_panel.csv already exists.')
elif os.path.exists('data'):
    file_out = 'data/data_panel.csv'
    data_panel.to_csv(path_or_buf=file_out, index=False)
    print('data/ already exists.')
    print('data/data_panel.csv has been created.')
else:
    os.mkdir('data')
    file_out = 'data/data_panel.csv'
    data_panel.to_csv(path_or_buf=file_out, index=False)
    print('data/data_panel.csv has been created.')
